+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Develop trunk changelog:
 
+- Adding optional plasticity settings to internal springs.
+- Correcting parameter lables/orderings in `SpringComponent`.
 - Adding method `Body.withComponent<T: BodyComponent>(ofType:, do:)` that fetches a body component of a specified type and passes it to a closure, if it exists within a body.
 - Fixing `Body.updateAABB()` w/ non-static bodies w/ `forceUpdate` flag set to `false`.
 - Fixing `World.update()` calling `Body.updateEdgesAndNormals()` one more time than necessary.